Renamed Facet component views
Renamed Facet component views so they have no indication of whether
they are for "Range" facets or "Value" facets. The facet components
should not care, they should be able to work with anything that has a
label, count, and value.

They should, however, indicate the type of input element they are,
and give an indication of whether or not they can be used with
multiple values or just a single value.

SingleRangeSelectFacet -> SingleSelectFacet
SingleValueLinksFacet -> SingleLinksFacet
MultiValueFacet -> MultiCheckboxFacet